+++ npm-search/src/js/app.js
@@ -1,6 +1,7 @@
-var React    = window.React = require('react'), // assign it to window for react chrome extension
-    classNames = require('classnames'),
-    App = {};
+var React           = window.React = require('react'), // assign it to window for react chrome extension
+    classNames      = require('classnames'),
+    SearchBar       = require('./searchbar.jsx'),
+    App             = {};
 
 var Constants = {
     DownloadCountBatch: 100,
@@ -102,31 +103,6 @@ var SupportedPlatforms = React.createClass({
                 {moreButton}
                 {otherPlatformsSupported}
             </ul>
-        );
-    }
-});
-
-var SearchBar = React.createClass({
-    handleChange: function() {
-        this.props.onUserInput(
-            this.refs.filterTextInput.getDOMNode().value
-        )
-    },
-    render: function() {
-        return (
-            <div className="col-xs-offset-2 col-xs-8">
-                <div id="searchwrapper">
-                    <input
-                        className="searchBox"
-                        type="search"
-                        autoComplete="off"
-                        placeholder={this.props.placeHolderText}
-                        value={this.props.filterText}
-                        ref="filterTextInput"
-                        onChange={this.handleChange}
-                    />
-                </div>
-            </div>
         );
     }
 });
